+++ wezterm-gui/src/termwindow/render/pane.rs
@@ -78,10 +78,6 @@ impl crate::TermWindow {
         let top_pixel_y = top_bar_height + padding_top + border.top.get() as f32;
 
         let cursor = pos.pane.get_cursor_position();
-        if pos.is_active {
-            self.prev_cursor.update(&cursor);
-        }
-
         let pane_id = pos.pane.pane_id();
         let current_viewport = self.get_viewport(pane_id);
         let dims = pos.pane.get_dimensions();